solved problem of floating dimensions

+++ problems/clp/objects2/Block.cpp
@@ -17,7 +17,7 @@ namespace clp {
 bool Block::FSB=false;
 set<const Block*, block_order> Block::all_blocks;
 
-Block::Block(long l, long w, long h) : Volume(l,w,h),occupied_volume(0), total_weight(0.0), total_profit(0.0), n_boxes(0),
+Block::Block(double l, double w, double h) : Volume(l,w,h),occupied_volume(0), total_weight(0.0), total_profit(0.0), n_boxes(0),
 		spaces(new SpaceSet(*this)), blocks(new AABBList()) {
 	spaces->insert(Space (*this));
 }
